--- Programming/.LaravelCore/app/Http/Controllers/Application/FrontEnd/SandBox/Examples_APICall/transaction/create/finance/setSalesInvoiceRequesitionDetail/v1/example.php
@@ -1,122 +1,0 @@
-<?php
-
-/*
-+----------------------------------------------------------------------------------------------------------------------------------+
-| ▪ Category    : Example - API Call Controller                                                                                    |
-| ▪ Name Space  : \App\Http\Controllers\Application\FrontEnd\SandBox\Examples_APICall\transaction\create\finance                   |
-|                 \setSalesInvoiceRequisitionDetail\v1                                                                             |
-| ▪ API Key     : transaction.create.finance.setSalesInvoiceRequisitionDetail                                                      |
-| ▪ API Version : 1                                                                                                                |
-|                                                                                                                                  |
-| ▪ Copyleft 🄯 2023 Zheta (teguhpjs@gmail.com)                                                                                     |
-+----------------------------------------------------------------------------------------------------------------------------------+
-*/
-namespace App\Http\Controllers\Application\FrontEnd\SandBox\Examples_APICall\transaction\create\finance\setSalesInvoiceRequisitionDetail\v1
-    {
-    class example extends \App\Http\Controllers\Controller
-        {
-        /*
-        +--------------------------------------------------------------------------------------------------------------------------+
-        | ▪ Call URL        : http(s)://<HOST>/transaction.create.finance.setSalesInvoiceRequisitionDetail.v1_throughAPIGateway    |
-        |                     ► http://172.28.0.4/transaction.create.finance.setSalesInvoiceRequisitionDetail.v1_throughAPIGateway |
-        +--------------------------------------------------------------------------------------------------------------------------+
-        | ▪ Version         : 1.0001.0000000                                                                                       |
-        | ▪ Last Update     : 2023-12-13                                                                                           |
-        | ▪ Creation Date   : 2023-11-13                                                                                           |
-        +--------------------------------------------------------------------------------------------------------------------------+
-        */
-        public function throughAPIGateway($varAPIWebToken)
-            {
-            //---Parameter Set---
-            if (!$varAPIWebToken) {
-                $varAPIWebToken = \App\Helpers\ZhtHelper\System\Helper_Environment::getUserSessionID_System();
-                }
-            //---Core---
-            $varData = \App\Helpers\ZhtHelper\System\FrontEnd\Helper_APICall::setCallAPIGateway(
-                \App\Helpers\ZhtHelper\System\Helper_Environment::getUserSessionID_System(),
-                $varAPIWebToken, 
-                'transaction.create.finance.setSalesInvoiceRequisitionDetail', 
-                'latest',
-                [
-                'entities' => [
-                    "salesInvoiceRequisition_RefID" => 244000000000001,
-                    'salesOrderDetail_RefID' => 130000000000001,
-                    "product_RefID" => 88000000000002,
-                    "quantity" => 10,
-                    "quantityUnit_RefID" => 73000000000001,
-                    'productUnitPriceCurrency_RefID' => 62000000000001,
-                    'productUnitPriceCurrencyValue' => 30000,
-                    'productUnitPriceCurrencyExchangeRate' => 1,
-                    'productUnitPriceDiscountCurrency_RefID' => 62000000000001,
-                    'productUnitPriceDiscountCurrencyValue' => 5000,
-                    'productUnitPriceDiscountCurrencyExchangeRate' => 1,
-                    "remarks" => 'Catatan'
-                    ]
-                ]
-                );
-            var_dump($varData);
-            }
-
-
-        /*
-        +--------------------------------------------------------------------------------------------------------------------------+
-        | ▪ Call URL        : http(s)://<HOST>/                                                                                    |
-        |                     transaction.create.finance.setSalesInvoiceRequisitionDetail.v1_throughAPIGatewayJQuery               |
-        |                     ► http://172.28.0.4/                                                                                 |
-        |                       transaction.create.finance.setSalesInvoiceRequisitionDetail.v1_throughAPIGatewayJQuery             |
-        +--------------------------------------------------------------------------------------------------------------------------+
-        | ▪ Version         : 1.0001.0000000                                                                                       |
-        | ▪ Last Update     : 2023-12-13                                                                                           |
-        | ▪ Creation Date   : 2023-11-13                                                                                           |
-        +--------------------------------------------------------------------------------------------------------------------------+
-        */
-        public function throughAPIGatewayJQuery($varAPIWebToken)
-            {
-            //---Parameter Set---
-            if (!$varAPIWebToken) {
-                $varAPIWebToken = \App\Helpers\ZhtHelper\System\Helper_Environment::getUserSessionID_System();
-                }
-            //---Core---
-            echo \App\Helpers\ZhtHelper\General\Helper_JavaScript::setLibrary(\App\Helpers\ZhtHelper\System\Helper_Environment::getUserSessionID_System());
-            echo '<table border="1" style="border-collapse: collapse;">';
-            echo    '<tr><td colspan="2" bgcolor="#6666cc" align="middle"><p style="color:#ffffff">Sales Invoice Requisition Detail Main Data</p></td></tr>';
-            echo       '<tr><td>SalesInvoiceRequisition_RefID</td><td><input type="text" id="dataInput_SalesInvoiceRequisition_RefID" value=244000000000001></td></tr>';
-            echo       '<tr><td>SalesOrderDetail_RefID</td><td><input type="text" id="dataInput_SalesOrderDetail_RefID" value=130000000000001></td></tr>';
-            echo       '<tr><td>Product_RefID</td><td><input type="text" id="dataInput_Product_RefID" value=88000000000002></td></tr>';
-            echo       '<tr><td>Quantity</td><td><input type="text" id="dataInput_Quantity" value=10></td></tr>';
-            echo       '<tr><td>QuantityUnit_RefID</td><td><input type="text" id="dataInput_QuantityUnit_RefID" value=73000000000001></td></tr>';
-            echo        '<tr><td>ProductUnitPriceCurrency_RefID</td><td><input type="text" id="dataInput_ProductUnitPriceCurrency_RefID" value=62000000000001></td></tr>';
-            echo        '<tr><td>ProductUnitPriceCurrencyValue</td><td><input type="text" id="dataInput_ProductUnitPriceCurrencyValue" value=30000></td></tr>';
-            echo        '<tr><td>ProductUnitPriceCurrencyExchangeRate</td><td><input type="text" id="dataInput_ProductUnitPriceCurrencyExchangeRate" value=1></td></tr>';
-            echo        '<tr><td>ProductUnitPriceDiscountCurrency_RefID</td><td><input type="text" id="dataInput_ProductUnitPriceDiscountCurrency_RefID" value=62000000000001></td></tr>';
-            echo        '<tr><td>ProductUnitPriceDiscountCurrencyValue</td><td><input type="text" id="dataInput_ProductUnitPriceDiscountCurrencyValue" value=5000></td></tr>';
-            echo        '<tr><td>ProductUnitPriceDiscountCurrencyExchangeRate</td><td><input type="text" id="dataInput_ProductUnitPriceDiscountCurrencyExchangeRate" value=1></td></tr>';
-            echo       '<tr><td>Remarks</td><td><input type="text" id="dataInput_Remarks" value="Catatan"></td></tr>';
-            echo '</table><br>';
-            $varJQueryFunction = \App\Helpers\ZhtHelper\System\FrontEnd\Helper_APICall::setCallAPIGatewayJQuery(
-                \App\Helpers\ZhtHelper\System\Helper_Environment::getUserSessionID_System(), 
-                $varAPIWebToken, 
-                'transaction.create.finance.setSalesInvoiceRequisitionDetail', 
-                'latest', 
-                '{'.
-                    '"entities" : {'.
-                        '"salesInvoiceRequisition_RefID" : parseInt(document.getElementById("dataInput_SalesInvoiceRequisition_RefID").value), '.
-                        '"salesOrderDetail_RefID" : parseInt(document.getElementById("dataInput_SalesOrderDetail_RefID").value), '.
-                        '"product_RefID" : parseInt(document.getElementById("dataInput_Product_RefID").value), '.
-                        '"quantity" : parseFloat(document.getElementById("dataInput_Quantity").value), '.
-                        '"quantityUnit_RefID" : parseInt(document.getElementById("dataInput_QuantityUnit_RefID").value), '.
-                        '"productUnitPriceCurrency_RefID" : parseInt(document.getElementById("dataInput_ProductUnitPriceCurrency_RefID").value), '.
-                        '"productUnitPriceCurrencyValue" : parseFloat(document.getElementById("dataInput_ProductUnitPriceCurrencyValue").value), '.
-                        '"productUnitPriceCurrencyExchangeRate" : parseFloat(document.getElementById("dataInput_ProductUnitPriceCurrencyExchangeRate").value), '.
-                        '"productUnitPriceDiscountCurrency_RefID" : parseInt(document.getElementById("dataInput_ProductUnitPriceDiscountCurrency_RefID").value), '.
-                        '"productUnitPriceDiscountCurrencyValue" : parseFloat(document.getElementById("dataInput_ProductUnitPriceDiscountCurrencyValue").value), '.
-                        '"productUnitPriceDiscountCurrencyExchangeRate" : parseFloat(document.getElementById("dataInput_ProductUnitPriceDiscountCurrencyExchangeRate").value), '.
-                        '"remarks" : document.getElementById("dataInput_Remarks").value'.
-                        '}'.
-                '}'
-                ); 
-            echo "<button type='button' onclick='javascript:var varData = ".$varJQueryFunction."; $(\"body\").append(JSON.stringify(varData));'>Submit Data</button>";
-            dd($varJQueryFunction);
-            }
-        }
-    }
