+++ Programming/.LaravelCore/app/Http/Controllers/Application/FrontEnd/SandBox/Examples_APICall/transaction/update/finance/setSalesInvoiceDetail/v1/example.php
@@ -8,7 +8,7 @@
 | ▪ API Key     : transaction.update.finance.setSalesInvoiceDetail                                                                 |
 | ▪ API Version : 1                                                                                                                |
 |                                                                                                                                  |
-| ▪ Copyleft 🄯 2023 Zheta (teguhpjs@gmail.com)                                                                                     |
+| ▪ Copyleft 🄯 2023 - 2025 Zheta (teguhpjs@gmail.com)                                                                              |
 +----------------------------------------------------------------------------------------------------------------------------------+
 */
 namespace App\Http\Controllers\Application\FrontEnd\SandBox\Examples_APICall\transaction\update\finance\setSalesInvoiceDetail\v1
@@ -20,8 +20,8 @@ namespace App\Http\Controllers\Application\FrontEnd\SandBox\Examples_APICall\tra
         | ▪ Call URL        : http(s)://<HOST>/transaction.update.finance.setSalesInvoiceDetail.v1_throughAPIGateway               |
         |                     ► http://172.28.0.4/transaction.update.finance.setSalesInvoiceDetail.v1_throughAPIGateway            |
         +--------------------------------------------------------------------------------------------------------------------------+
-        | ▪ Version         : 1.0001.0000000                                                                                       |
-        | ▪ Last Update     : 2023-12-12                                                                                           |
+        | ▪ Version         : 1.0002.0000000                                                                                       |
+        | ▪ Last Update     : 2025-01-23                                                                                           |
         | ▪ Creation Date   : 2023-11-21                                                                                           |
         +--------------------------------------------------------------------------------------------------------------------------+
         */
@@ -29,58 +29,67 @@ namespace App\Http\Controllers\Application\FrontEnd\SandBox\Examples_APICall\tra
             {
             //---Parameter Set---
             if (!$varAPIWebToken) {
-                $varAPIWebToken = \App\Helpers\ZhtHelper\System\Helper_Environment::getUserSessionID_System();
+                $varAPIWebToken =
+                    \App\Helpers\ZhtHelper\System\Helper_Environment::getAPIWebToken_System();
                 }
+
             //---Core---
-            $varData = \App\Helpers\ZhtHelper\System\FrontEnd\Helper_APICall::setCallAPIGateway(
-                \App\Helpers\ZhtHelper\System\Helper_Environment::getUserSessionID_System(),
-                $varAPIWebToken, 
-                'transaction.update.finance.setSalesInvoiceDetail', 
-                'latest',
-                [
-                "recordID" => 139000000000001,
-                "entities" => [
-                    "salesInvoice_RefID" => 138000000000001,
-                    "salesInvoiceRequisitionDetail_RefID" => 245000000000001,
-                    "product_RefID" => 88000000000002,
-                    "quantity" => 10,
-                    "quantityUnit_RefID" => 73000000000001,
-                    "productUnitPriceCurrency_RefID" => 62000000000001,
-                    "productUnitPriceCurrencyValue" => 30000,
-                    "productUnitPriceCurrencyExchangeRate" => 1,
-                    "productUnitPriceDiscountCurrency_RefID" => 62000000000001,
-                    "productUnitPriceDiscountCurrencyValue" => 5000,
-                    "productUnitPriceDiscountCurrencyExchangeRate" => 1,
-                    "remarks" => 'Catatan',
-                    "additionalData" => [
-                        "transactionTaxItemList" => [
-                            "items" => [
-                                    [
-                                    "recordID" => 185000000000001,
-                                    "entities" => [
-                                        "taxType_RefID" => 182000000000001,
-                                        "tariffCurrency_RefID" => 62000000000001,
-                                        "tariffCurrencyValue" => 3000,
-                                        "tariffCurrencyExchangeRate" => 1,
-                                        "remarks" => null
-                                        ]
-                                    ],
-                                    [
-                                    "recordID" => 185000000000002,
-                                    "entities" => [
-                                        "taxType_RefID" => 182000000000005,
-                                        "tariffCurrency_RefID" => 62000000000001,
-                                        "tariffCurrencyValue" => 4000,
-                                        "tariffCurrencyExchangeRate" => 1,
-                                        "remarks" => null
+            $varData =
+                \App\Helpers\ZhtHelper\System\FrontEnd\Helper_APICall::setCallAPIGateway(
+                    //-----[ HEADER ]-----( START )-----
+                        \App\Helpers\ZhtHelper\System\Helper_Environment::getUserSessionID_System(),
+                        $varAPIWebToken, 
+                        'transaction.update.finance.setSalesInvoiceDetail', 
+                        'latest',
+                    //-----[ HEADER ]-----( END )-----
+
+                    //-----[ BODY ]-----( START )-----
+                        [
+                        "recordID" => 139000000000001,
+                        "entities" => [
+                            "salesInvoice_RefID" => 138000000000001,
+                            "salesInvoiceRequisitionDetail_RefID" => 245000000000001,
+                            "product_RefID" => 88000000000002,
+                            "quantity" => 10,
+                            "quantityUnit_RefID" => 73000000000001,
+                            "productUnitPriceCurrency_RefID" => 62000000000001,
+                            "productUnitPriceCurrencyValue" => 30000,
+                            "productUnitPriceCurrencyExchangeRate" => 1,
+                            "productUnitPriceDiscountCurrency_RefID" => 62000000000001,
+                            "productUnitPriceDiscountCurrencyValue" => 5000,
+                            "productUnitPriceDiscountCurrencyExchangeRate" => 1,
+                            "remarks" => 'Catatan',
+                            "additionalData" => [
+                                "transactionTaxItemList" => [
+                                    "items" => [
+                                            [
+                                            "recordID" => 185000000000001,
+                                            "entities" => [
+                                                "taxType_RefID" => 182000000000001,
+                                                "tariffCurrency_RefID" => 62000000000001,
+                                                "tariffCurrencyValue" => 3000,
+                                                "tariffCurrencyExchangeRate" => 1,
+                                                "remarks" => null
+                                                ]
+                                            ],
+                                            [
+                                            "recordID" => 185000000000002,
+                                            "entities" => [
+                                                "taxType_RefID" => 182000000000005,
+                                                "tariffCurrency_RefID" => 62000000000001,
+                                                "tariffCurrencyValue" => 4000,
+                                                "tariffCurrencyExchangeRate" => 1,
+                                                "remarks" => null
+                                                ]
+                                            ]
                                         ]
                                     ]
                                 ]
                             ]
                         ]
-                    ]
-                ]
-                );
+                    //-----[ BODY ]-----( END )-----
+                    );
+
             var_dump($varData);
             }
 
@@ -91,8 +100,8 @@ namespace App\Http\Controllers\Application\FrontEnd\SandBox\Examples_APICall\tra
         |                     ► http://172.28.0.4/                                                                                 |
         |                       transaction.update.finance.setSalesInvoiceDetail.v1_throughAPIGatewayJQuery                        |
         +--------------------------------------------------------------------------------------------------------------------------+
-        | ▪ Version         : 1.0001.0000000                                                                                       |
-        | ▪ Last Update     : 2023-12-12                                                                                           |
+        | ▪ Version         : 1.0002.0000000                                                                                       |
+        | ▪ Last Update     : 2025-01-23                                                                                           |
         | ▪ Creation Date   : 2023-11-21                                                                                           |
         +--------------------------------------------------------------------------------------------------------------------------+
         */
@@ -100,10 +109,13 @@ namespace App\Http\Controllers\Application\FrontEnd\SandBox\Examples_APICall\tra
             {
             //---Parameter Set---
             if (!$varAPIWebToken) {
-                $varAPIWebToken = \App\Helpers\ZhtHelper\System\Helper_Environment::getUserSessionID_System();
+                $varAPIWebToken =
+                    \App\Helpers\ZhtHelper\System\Helper_Environment::getAPIWebToken_System();
                 }
+
             //---Core---
             echo \App\Helpers\ZhtHelper\General\Helper_JavaScript::setLibrary(\App\Helpers\ZhtHelper\System\Helper_Environment::getUserSessionID_System());
+
             echo '<table border="1" style="border-collapse: collapse;">';
             echo    '<tr><td colspan="2" bgcolor="#6666cc" align="middle"><p style="color:#ffffff">Sales Invoice Detail Main Data</p></td></tr>';
             echo        '<tr><td>RecordID</td><td><input type="text" id="dataInput_RecordID" value=139000000000001></td></tr>';
@@ -135,59 +147,66 @@ namespace App\Http\Controllers\Application\FrontEnd\SandBox\Examples_APICall\tra
             echo        '<tr><td>Remarks_2</td><td><input type="text" id="dataInput_Remarks_2" value="Catatan Kedua"></td></tr>';
             echo '</table><br>';
 
-            $varJQueryFunction = \App\Helpers\ZhtHelper\System\FrontEnd\Helper_APICall::setCallAPIGatewayJQuery(
-                \App\Helpers\ZhtHelper\System\Helper_Environment::getUserSessionID_System(), 
-                $varAPIWebToken, 
-                'transaction.update.finance.setSalesInvoiceDetail', 
-                'latest', 
-                '{'.
-                    '"recordID" : parseInt(document.getElementById("dataInput_RecordID").value), '.
-                    '"entities" : {'.
-                        '"salesInvoice_RefID" : parseInt(document.getElementById("dataInput_SalesInvoice_RefID").value), '.
-                        '"salesInvoiceRequisitionDetail_RefID" : parseInt(document.getElementById("dataInput_SalesInvoiceRequisitionDetail_RefID").value), '.
-                        '"product_RefID" : parseInt(document.getElementById("dataInput_Product_RefID").value), '.
-                        '"quantity" : parseFloat(document.getElementById("dataInput_Quantity").value), '.
-                        '"quantityUnit_RefID" : parseInt(document.getElementById("dataInput_QuantityUnit_RefID").value), '.
-                        '"productUnitPriceCurrency_RefID" : parseInt(document.getElementById("dataInput_ProductUnitPriceCurrency_RefID").value), '.
-                        '"productUnitPriceCurrencyValue" : parseFloat(document.getElementById("dataInput_ProductUnitPriceCurrencyValue").value), '.
-                        '"productUnitPriceCurrencyExchangeRate" : parseFloat(document.getElementById("dataInput_ProductUnitPriceCurrencyExchangeRate").value), '.
-                        '"productUnitPriceDiscountCurrency_RefID" : parseInt(document.getElementById("dataInput_ProductUnitPriceDiscountCurrency_RefID").value), '.
-                        '"productUnitPriceDiscountCurrencyValue" : parseFloat(document.getElementById("dataInput_ProductUnitPriceDiscountCurrencyValue").value), '.
-                        '"productUnitPriceDiscountCurrencyExchangeRate" : parseFloat(document.getElementById("dataInput_ProductUnitPriceDiscountCurrencyExchangeRate").value), '.
-                        '"remarks" : document.getElementById("dataInput_Remarks").value, '.
-                        '"additionalData" : {'.
-                            '"transactionTaxItemList" : {'.
-                                '"items" : ['.
-                                        '{'.
-                                        '"recordID" : parseInt(document.getElementById("dataInput_RecordIDDetail_RefID_1").value), '.
-                                        '"entities" : '.
+            $varJQueryFunction =
+                \App\Helpers\ZhtHelper\System\FrontEnd\Helper_APICall::setCallAPIGatewayJQuery(
+                    //-----[ HEADER ]-----( START )-----
+                        \App\Helpers\ZhtHelper\System\Helper_Environment::getUserSessionID_System(), 
+                        $varAPIWebToken, 
+                        'transaction.update.finance.setSalesInvoiceDetail', 
+                        'latest',
+                    //-----[ HEADER ]-----( END )-----
+
+                    //-----[ BODY ]-----( START )-----
+                        '{'.
+                        '"recordID" : parseInt(document.getElementById("dataInput_RecordID").value), '.
+                        '"entities" : {'.
+                            '"salesInvoice_RefID" : parseInt(document.getElementById("dataInput_SalesInvoice_RefID").value), '.
+                            '"salesInvoiceRequisitionDetail_RefID" : parseInt(document.getElementById("dataInput_SalesInvoiceRequisitionDetail_RefID").value), '.
+                            '"product_RefID" : parseInt(document.getElementById("dataInput_Product_RefID").value), '.
+                            '"quantity" : parseFloat(document.getElementById("dataInput_Quantity").value), '.
+                            '"quantityUnit_RefID" : parseInt(document.getElementById("dataInput_QuantityUnit_RefID").value), '.
+                            '"productUnitPriceCurrency_RefID" : parseInt(document.getElementById("dataInput_ProductUnitPriceCurrency_RefID").value), '.
+                            '"productUnitPriceCurrencyValue" : parseFloat(document.getElementById("dataInput_ProductUnitPriceCurrencyValue").value), '.
+                            '"productUnitPriceCurrencyExchangeRate" : parseFloat(document.getElementById("dataInput_ProductUnitPriceCurrencyExchangeRate").value), '.
+                            '"productUnitPriceDiscountCurrency_RefID" : parseInt(document.getElementById("dataInput_ProductUnitPriceDiscountCurrency_RefID").value), '.
+                            '"productUnitPriceDiscountCurrencyValue" : parseFloat(document.getElementById("dataInput_ProductUnitPriceDiscountCurrencyValue").value), '.
+                            '"productUnitPriceDiscountCurrencyExchangeRate" : parseFloat(document.getElementById("dataInput_ProductUnitPriceDiscountCurrencyExchangeRate").value), '.
+                            '"remarks" : document.getElementById("dataInput_Remarks").value, '.
+                            '"additionalData" : {'.
+                                '"transactionTaxItemList" : {'.
+                                    '"items" : ['.
                                             '{'.
-                                            '"taxType_RefID" : parseInt(document.getElementById("dataInput_TaxType_RefID_1").value), '.
-                                            '"tariffCurrency_RefID" : parseInt(document.getElementById("dataInput_TariffCurrency_RefID_1").value), '.
-                                            '"tariffCurrencyValue" : parseFloat(document.getElementById("dataInput_TariffCurrencyValue_1").value), '.
-                                            '"tariffCurrencyExchangeRate" : parseFloat(document.getElementById("dataInput_TariffCurrencyExchangeRate_1").value), '.
-                                            '"remarks" : document.getElementById("dataInput_Remarks_1").value'.
-                                            '}'.
-                                        '}, '.
-                                        '{'.
-                                        '"recordID" : parseInt(document.getElementById("dataInput_RecordIDDetail_RefID_2").value), '.
-                                        '"entities" : '.
+                                            '"recordID" : parseInt(document.getElementById("dataInput_RecordIDDetail_RefID_1").value), '.
+                                            '"entities" : '.
+                                                '{'.
+                                                '"taxType_RefID" : parseInt(document.getElementById("dataInput_TaxType_RefID_1").value), '.
+                                                '"tariffCurrency_RefID" : parseInt(document.getElementById("dataInput_TariffCurrency_RefID_1").value), '.
+                                                '"tariffCurrencyValue" : parseFloat(document.getElementById("dataInput_TariffCurrencyValue_1").value), '.
+                                                '"tariffCurrencyExchangeRate" : parseFloat(document.getElementById("dataInput_TariffCurrencyExchangeRate_1").value), '.
+                                                '"remarks" : document.getElementById("dataInput_Remarks_1").value'.
+                                                '}'.
+                                            '}, '.
                                             '{'.
-                                            '"taxType_RefID" : parseInt(document.getElementById("dataInput_TaxType_RefID_2").value), '.
-                                            '"tariffCurrency_RefID" : parseInt(document.getElementById("dataInput_TariffCurrency_RefID_2").value), '.
-                                            '"tariffCurrencyValue" : parseFloat(document.getElementById("dataInput_TariffCurrencyValue_2").value), '.
-                                            '"tariffCurrencyExchangeRate" : parseFloat(document.getElementById("dataInput_TariffCurrencyExchangeRate_2").value), '.
-                                            '"remarks" : document.getElementById("dataInput_Remarks_2").value'.
+                                            '"recordID" : parseInt(document.getElementById("dataInput_RecordIDDetail_RefID_2").value), '.
+                                            '"entities" : '.
+                                                '{'.
+                                                '"taxType_RefID" : parseInt(document.getElementById("dataInput_TaxType_RefID_2").value), '.
+                                                '"tariffCurrency_RefID" : parseInt(document.getElementById("dataInput_TariffCurrency_RefID_2").value), '.
+                                                '"tariffCurrencyValue" : parseFloat(document.getElementById("dataInput_TariffCurrencyValue_2").value), '.
+                                                '"tariffCurrencyExchangeRate" : parseFloat(document.getElementById("dataInput_TariffCurrencyExchangeRate_2").value), '.
+                                                '"remarks" : document.getElementById("dataInput_Remarks_2").value'.
+                                                '}'.
                                             '}'.
-                                        '}'.
-                                    ']'.
+                                        ']'.
+                                    '}'.
                                 '}'.
                             '}'.
-                        '}'.
+                        '}'
+                    //-----[ BODY ]-----( END )-----
+                    );
 
-                '}'
-                ); 
             echo "<button type='button' onclick='javascript:var varData = ".$varJQueryFunction."; $(\"body\").append(JSON.stringify(varData));'>Submit Data</button>";
+
             dd($varJQueryFunction);
             }
         }
