+++ Programming/.LaravelCore/app/Http/Controllers/Application/FrontEnd/SandBox/Examples_APICall/transaction/update/finance/setSalesInvoiceRequisitionDetail/v1/example.php
@@ -8,7 +8,7 @@
 | ▪ API Key     : transaction.update.finance.setSalesInvoiceRequisitionDetail                                                      |
 | ▪ API Version : 1                                                                                                                |
 |                                                                                                                                  |
-| ▪ Copyleft 🄯 2023 Zheta (teguhpjs@gmail.com)                                                                                     |
+| ▪ Copyleft 🄯 2023 - 2025 Zheta (teguhpjs@gmail.com)                                                                              |
 +----------------------------------------------------------------------------------------------------------------------------------+
 */
 namespace App\Http\Controllers\Application\FrontEnd\SandBox\Examples_APICall\transaction\update\finance\setSalesInvoiceRequisitionDetail\v1
@@ -20,8 +20,8 @@ namespace App\Http\Controllers\Application\FrontEnd\SandBox\Examples_APICall\tra
         | ▪ Call URL        : http(s)://<HOST>/transaction.update.finance.setSalesInvoiceRequisitionDetail.v1_throughAPIGateway    |
         |                     ► http://172.28.0.4/transaction.update.finance.setSalesInvoiceRequisitionDetail.v1_throughAPIGateway |
         +--------------------------------------------------------------------------------------------------------------------------+
-        | ▪ Version         : 1.0001.0000000                                                                                       |
-        | ▪ Last Update     : 2023-12-12                                                                                           |
+        | ▪ Version         : 1.0002.0000000                                                                                       |
+        | ▪ Last Update     : 2025-01-23                                                                                           |
         | ▪ Creation Date   : 2023-11-13                                                                                           |
         +--------------------------------------------------------------------------------------------------------------------------+
         */
@@ -29,32 +29,41 @@ namespace App\Http\Controllers\Application\FrontEnd\SandBox\Examples_APICall\tra
             {
             //---Parameter Set---
             if (!$varAPIWebToken) {
-                $varAPIWebToken = \App\Helpers\ZhtHelper\System\Helper_Environment::getUserSessionID_System();
+                $varAPIWebToken =
+                    \App\Helpers\ZhtHelper\System\Helper_Environment::getAPIWebToken_System();
                 }
+
             //---Core---
-            $varData = \App\Helpers\ZhtHelper\System\FrontEnd\Helper_APICall::setCallAPIGateway(
-                \App\Helpers\ZhtHelper\System\Helper_Environment::getUserSessionID_System(),
-                $varAPIWebToken, 
-                'transaction.update.finance.setSalesInvoiceRequisitionDetail', 
-                'latest', 
-                [
-                'recordID' => 245000000000001,
-                'entities' => [
-                    'salesInvoiceRequisition_RefID' => 244000000000001,
-                    'salesOrderDetail_RefID' => 130000000000001,
-                    'product_RefID' => 88000000000002,
-                    'quantity' => 10,
-                    'quantityUnit_RefID' => 73000000000001,
-                    'productUnitPriceCurrency_RefID' => 62000000000001,
-                    'productUnitPriceCurrencyValue' => 30000,
-                    'productUnitPriceCurrencyExchangeRate' => 1,
-                    'productUnitPriceDiscountCurrency_RefID' => 62000000000001,
-                    'productUnitPriceDiscountCurrencyValue' => 5000,
-                    'productUnitPriceDiscountCurrencyExchangeRate' => 1,
-                    'remarks' => 'Catatan'
-                    ]
-                ]
-                );
+            $varData =
+                \App\Helpers\ZhtHelper\System\FrontEnd\Helper_APICall::setCallAPIGateway(
+                    //-----[ HEADER ]-----( START )-----
+                        \App\Helpers\ZhtHelper\System\Helper_Environment::getUserSessionID_System(),
+                        $varAPIWebToken, 
+                        'transaction.update.finance.setSalesInvoiceRequisitionDetail', 
+                        'latest',
+                    //-----[ HEADER ]-----( END )-----
+
+                    //-----[ BODY ]-----( START )-----
+                        [
+                        'recordID' => 245000000000001,
+                        'entities' => [
+                            'salesInvoiceRequisition_RefID' => 244000000000001,
+                            'salesOrderDetail_RefID' => 130000000000001,
+                            'product_RefID' => 88000000000002,
+                            'quantity' => 10,
+                            'quantityUnit_RefID' => 73000000000001,
+                            'productUnitPriceCurrency_RefID' => 62000000000001,
+                            'productUnitPriceCurrencyValue' => 30000,
+                            'productUnitPriceCurrencyExchangeRate' => 1,
+                            'productUnitPriceDiscountCurrency_RefID' => 62000000000001,
+                            'productUnitPriceDiscountCurrencyValue' => 5000,
+                            'productUnitPriceDiscountCurrencyExchangeRate' => 1,
+                            'remarks' => 'Catatan'
+                            ]
+                        ]
+                    //-----[ BODY ]-----( END )-----
+                    );
+
             var_dump($varData);
             }
 
@@ -66,8 +75,8 @@ namespace App\Http\Controllers\Application\FrontEnd\SandBox\Examples_APICall\tra
         |                     ► http://172.28.0.4/                                                                                 |
         |                       transaction.update.finance.setSalesInvoiceRequisitionDetail.v1_throughAPIGatewayJQuery             |
         +--------------------------------------------------------------------------------------------------------------------------+
-        | ▪ Version         : 1.0001.0000000                                                                                       |
-        | ▪ Last Update     : 2023-12-12                                                                                           |
+        | ▪ Version         : 1.0002.0000000                                                                                       |
+        | ▪ Last Update     : 2025-01-23                                                                                           |
         | ▪ Creation Date   : 2023-11-13                                                                                           |
         +--------------------------------------------------------------------------------------------------------------------------+
         */
@@ -75,10 +84,13 @@ namespace App\Http\Controllers\Application\FrontEnd\SandBox\Examples_APICall\tra
             {
             //---Parameter Set---
             if (!$varAPIWebToken) {
-                $varAPIWebToken = \App\Helpers\ZhtHelper\System\Helper_Environment::getUserSessionID_System();
+                $varAPIWebToken =
+                    \App\Helpers\ZhtHelper\System\Helper_Environment::getAPIWebToken_System();
                 }
+
             //---Core---
             echo \App\Helpers\ZhtHelper\General\Helper_JavaScript::setLibrary(\App\Helpers\ZhtHelper\System\Helper_Environment::getUserSessionID_System());
+
             echo '<table border="1" style="border-collapse: collapse;">';
             echo    '<tr><td colspan="2" bgcolor="#6666cc" align="middle"><p style="color:#ffffff">Sales Invoice Requisition Detail Main Data</p></td></tr>';
             echo        '<tr><td>RecordID</td><td><input type="text" id="dataInput_RecordID" value=245000000000001></td></tr>';
@@ -95,30 +107,39 @@ namespace App\Http\Controllers\Application\FrontEnd\SandBox\Examples_APICall\tra
             echo        '<tr><td>ProductUnitPriceDiscountCurrencyExchangeRate</td><td><input type="text" id="dataInput_ProductUnitPriceDiscountCurrencyExchangeRate" value=1></td></tr>';
             echo        '<tr><td>Remarks</td><td><input type="text" id="dataInput_Remarks" value="Catatan"></td></tr>';
             echo '</table><br>';
-            $varJQueryFunction = \App\Helpers\ZhtHelper\System\FrontEnd\Helper_APICall::setCallAPIGatewayJQuery(
-                \App\Helpers\ZhtHelper\System\Helper_Environment::getUserSessionID_System(), 
-                $varAPIWebToken, 
-                'transaction.update.finance.setSalesInvoiceRequisitionDetail', 
-                'latest', 
-                '{'.
-                    '"recordID" : parseInt(document.getElementById("dataInput_RecordID").value), '.
-                    '"entities" : {'.
-                        '"salesInvoiceRequisition_RefID" : parseInt(document.getElementById("dataInput_SalesInvoiceRequisition_RefID").value), '.
-                        '"salesOrderDetail_RefID" : parseInt(document.getElementById("dataInput_SalesOrderDetail_RefID").value), '.
-                        '"product_RefID" : parseInt(document.getElementById("dataInput_Product_RefID").value), '.
-                        '"quantity" : parseFloat(document.getElementById("dataInput_Quantity").value), '.
-                        '"quantityUnit_RefID" : parseInt(document.getElementById("dataInput_QuantityUnit_RefID").value), '.
-                        '"productUnitPriceCurrency_RefID" : parseInt(document.getElementById("dataInput_ProductUnitPriceCurrency_RefID").value), '.
-                        '"productUnitPriceCurrencyValue" : parseFloat(document.getElementById("dataInput_ProductUnitPriceCurrencyValue").value), '.
-                        '"productUnitPriceCurrencyExchangeRate" : parseFloat(document.getElementById("dataInput_ProductUnitPriceCurrencyExchangeRate").value), '.
-                        '"productUnitPriceDiscountCurrency_RefID" : parseInt(document.getElementById("dataInput_ProductUnitPriceDiscountCurrency_RefID").value), '.
-                        '"productUnitPriceDiscountCurrencyValue" : parseFloat(document.getElementById("dataInput_ProductUnitPriceDiscountCurrencyValue").value), '.
-                        '"productUnitPriceDiscountCurrencyExchangeRate" : parseFloat(document.getElementById("dataInput_ProductUnitPriceDiscountCurrencyExchangeRate").value), '.
-                        '"remarks" : document.getElementById("dataInput_Remarks").value'.
-                        '}'.
-                '}'
-                ); 
+
+            $varJQueryFunction =
+                \App\Helpers\ZhtHelper\System\FrontEnd\Helper_APICall::setCallAPIGatewayJQuery(
+                    //-----[ HEADER ]-----( START )-----
+                        \App\Helpers\ZhtHelper\System\Helper_Environment::getUserSessionID_System(), 
+                        $varAPIWebToken, 
+                        'transaction.update.finance.setSalesInvoiceRequisitionDetail', 
+                        'latest',
+                    //-----[ HEADER ]-----( END )-----
+
+                    //-----[ BODY ]-----( START )-----
+                        '{'.
+                        '"recordID" : parseInt(document.getElementById("dataInput_RecordID").value), '.
+                        '"entities" : {'.
+                            '"salesInvoiceRequisition_RefID" : parseInt(document.getElementById("dataInput_SalesInvoiceRequisition_RefID").value), '.
+                            '"salesOrderDetail_RefID" : parseInt(document.getElementById("dataInput_SalesOrderDetail_RefID").value), '.
+                            '"product_RefID" : parseInt(document.getElementById("dataInput_Product_RefID").value), '.
+                            '"quantity" : parseFloat(document.getElementById("dataInput_Quantity").value), '.
+                            '"quantityUnit_RefID" : parseInt(document.getElementById("dataInput_QuantityUnit_RefID").value), '.
+                            '"productUnitPriceCurrency_RefID" : parseInt(document.getElementById("dataInput_ProductUnitPriceCurrency_RefID").value), '.
+                            '"productUnitPriceCurrencyValue" : parseFloat(document.getElementById("dataInput_ProductUnitPriceCurrencyValue").value), '.
+                            '"productUnitPriceCurrencyExchangeRate" : parseFloat(document.getElementById("dataInput_ProductUnitPriceCurrencyExchangeRate").value), '.
+                            '"productUnitPriceDiscountCurrency_RefID" : parseInt(document.getElementById("dataInput_ProductUnitPriceDiscountCurrency_RefID").value), '.
+                            '"productUnitPriceDiscountCurrencyValue" : parseFloat(document.getElementById("dataInput_ProductUnitPriceDiscountCurrencyValue").value), '.
+                            '"productUnitPriceDiscountCurrencyExchangeRate" : parseFloat(document.getElementById("dataInput_ProductUnitPriceDiscountCurrencyExchangeRate").value), '.
+                            '"remarks" : document.getElementById("dataInput_Remarks").value'.
+                            '}'.
+                        '}'
+                    //-----[ BODY ]-----( END )-----
+                    ); 
+
             echo "<button type='button' onclick='javascript:var varData = ".$varJQueryFunction."; $(\"body\").append(JSON.stringify(varData));'>Submit Data</button>";
+
             dd($varJQueryFunction);
             }
         }
